api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
torch==2.1.0
numpy==1.26.2
pandas==2.1.3
pydantic==2.5.0
python-multipart==0.0.6
